poslength query working

+++ hu.bme.mit.incqueryd/hu.bme.mit.incqueryd.core/src/main/java/hu/bme/mit/incqueryd/rete/comparison/ConditionExpression.java
@@ -20,7 +20,8 @@ public class ConditionExpression {
     }
 
     public boolean satisfiesCondition(final Tuple tuple) {
-        final Long tupleField = (Long) tuple.get(index);
+//        final Long tupleField = (Long) tuple.get(index);
+    	final Integer tupleField = (Integer) tuple.get(index);
         return Compare.compare(tupleField, operand, comparisonOperator);
     }
